requirements.txt
confluent-kafka
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
faker==19.13.0
pydantic==2.5.0
python-dateutil==2.8.2
paho-mqtt==1.6.1
PyJWT==2.8.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pandas==2.1.4
numpy==1.26.2
prophet==1.1.5
scikit-learn==1.3.2
xgboost==2.0.3
torch==2.1.2
joblib==1.3.2
matplotlib==3.8.2
uvicorn==0.24.0
fastapi==0.104.1
